rely on xray

+++ lib/backtracer_ping.rb
@@ -1,4 +1,10 @@
 require 'pp'
+require 'sane/os'
+
+def xray
+ require 'xray'
+ proc { Process.kill "QUIT", Process.pid; '' }
+end
 
 if Thread.current.respond_to? :backtrace
   fella = proc { 
@@ -9,13 +15,23 @@ if Thread.current.respond_to? :backtrace
     out    
   }
 elsif respond_to? :caller_for_all_threads
-  fella = proc {
-    caller_for_all_threads
-  }
+  if OS.windows?
+    fella = proc {
+      caller_for_all_threads
+    }
+  else
+    fella = xray
+  end
 else
  # weak sauce for the old school users :)
- trap("ILL") { puts "All threads:" + Thread.list.inspect, "Current thread:" + Thread.current.to_s, caller } # puts current thread caller
- fella = proc {  Process.kill "ILL", Process.pid } # send myself a signal
+ if OS.windows?
+
+   trap("ILL") { puts "All threads:" + Thread.list.inspect, "Current thread:" + Thread.current.to_s, caller } # puts current thread caller
+   fella = proc {  Process.kill "ILL", Process.pid } # send myself a signal
+ else
+   fella = xray
+ end
+
 end
 
 time = $ping_interval || 5 # seconds
